The tf_files have the dataset(i have not uploaded), models, test images.
The scripts folder have various files used for training and testing the model. The retrain.py is used to start the training, the default values of training steps and learning rate had been changed to certain value. The output model files are uploaded on drive, due to size constraints of github.